requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
black==23.7.0
celery==5.3.1
certifi==2024.6.2
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
Django==4.2
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-filter==23.2
django-ipware==7.0.1
django-js-asset==2.1.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-redis==5.4.0
django-request==1.6.3
django-rest-framework==0.1.0
django-seed==0.2.2
django-simple-captcha==0.5.20
django-timezone-field==6.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-recaptcha==3.1.0
drf-yasg==1.21.7
Faker==26.0.0
generics==6.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonify==0.5
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.1
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.7
pycodestyle==2.11.0
PyJWT==1.7.1
python-crontab==3.1.0
python-dateutil==2.8.2
python-dotenv==0.15.0
python-ipware==3.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.25.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
